Added bottomTabs
